Moves break inside store.is_key_entry logic
This PR was in response to an issue we were running into when converting
Apple .p12 files into .pem using `OpenSSL::PKCS12`.

When creating a new instance with an Apple .p12 file, the initializer
was only checking the first alias in the `store` because of a break in the
.each loop. If there is more than one alias, only the first is being
checked.

In the case of the our .p12 file, the SECOND alias passes
`store.is_key_store`, but was being skipped because of the break.

This moves the break inside the `store.is_key_entry` logic statement,
which means each alias will be iterated through until either a match is
found or there are no other aliases.