Instantiate both the private and public keys when setting parameters.
The rsa_n and rsa_e parameter fields are used to instantiate both the
private and public keys, but are nulled after use in
generatePrivateKeyIfParams() and generatePublicKeyIfParams(). This meant
that once one of the keys had been created, the other could not be. The
RSA object could not then be used for both private and public functions.

If a private or public key has already been instantiated, use the key as
the source for the n and e parameters in generatePrivateKeyIfParams()
and generatePublicKeyIfParams().