Fix ClassCastException in X509Store.verify
This commit changes the logic of the `X509StoreContext.initialize`
method, to address some test failures that were showing up in
`test_x509store.rb`.  The failures would occur during calls to
the `verify` method of `X509Store`.

It seems that the failures were introduced by the addition of the
`X509Cert.toJava` method in ecc1a05f458bdc05d6d0f6fd025f8788ded0f120 .
The `toJava` method seems to be implicitly called when iterating over
a RubyArray in Java code, and the implementation of that method
now returns a BC object rather than the expected JRuby `X509Cert`
object.

This commit changes the iteration to use the `.eltOk` method to access
the underlying objects w/o the implicit call to `toJava`.