Try direct path for MessageDigest before invasive path
This logic is responsible for the remaining known "illegal access"
warning on Java 9+. I am not sure why we do not trust the default
provide to give us most algorithms, since it should be identical
to the one provied by Bouncy Castle. This change reverses the
logic and only uses the invasive provider-specific form when the
direct lookup fails.

Relates to jruby/jruby#6098.