Add failing test for verify_certificate_identity
This certificate and key is loaded from a fixtures file as I couldn't work out
how to create an example programatically.

The certificate has a `subjectAltName` extension with an `otherName` entry.

`verify_certificate_identity` fails because it parses the string representation
of `subjectAltName` with a simple regex expecting entries in the sequence to be
comma separated only.